Implement Factory Method and Command design patterns, refactor and add more tests

+++ Microservices/ChatService/src/main/java/com/example/chatservice/services/MessageService.java
@@ -2,10 +2,14 @@ package com.example.chatservice.services;
 
 import com.example.chatservice.models.Message;
 import com.example.chatservice.repositories.MessageRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.UUID;
+
+import com.example.chatservice.enums.MessageStatus;
 
 @Service
 public class MessageService {
@@ -29,6 +33,22 @@ public class MessageService {
     }
 
     public void deleteMessage(UUID id) {
+
+        Message message = getMessageById(id);
+        if (message == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Message not found");
+        }
+
         messageRepository.deleteById(id);
+    }
+
+    public boolean isMessageSeen(UUID messageId) {
+        Message message = getMessageById(messageId);
+
+        if (message == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Message not found");
+        }
+
+        return message.getStatus() == MessageStatus.SEEN;
     }
 }
